pyproject.toml
[tool.poetry]
name = "LiteSecBot"
version = "1.0.0"
description = "A Discord Bot Made by Raul Victor"
authors = ["Raul Victor <victorraulsm.developer@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11.6"
python-dotenv = "^0.21.0"
PyYAML = "^6.0"
PyNaCl = "^1.5.0"
Markdown = "^3.3.7"
discord-webhook = "^0.17.0"
Jinja2 = "^3.1.2"
nextcord = {extras = ["speed"], version = "^2.2.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"